+++ jest.setup.ts
@@ -1,0 +1,2 @@
+// Increase timeout (due to async communication with real backend services)
+jasmine.DEFAULT_TIMEOUT_INTERVAL = 100000;
