+++ test/interfaces/git-conventional-commit.interface.ts
@@ -1,0 +1,9 @@
+/**
+ * Git Conventional Commit interface
+ */
+export interface GitConventionalCommit {
+	type?: string;
+	scope?: string;
+	message: string;
+	hash?: string;
+}
