+++ test/interfaces/prepared-commit.interface.ts
@@ -1,0 +1,10 @@
+import { GitConventionalCommit } from './git-conventional-commit.interface';
+
+/**
+ * Prepared commit
+ */
+export interface PreparedCommit {
+    fileName: string;
+    fileContent: string;
+    commit: GitConventionalCommit;
+}
